Float conversion fixes + some more conversions

+++ test/mppa/instr/fnegw.c
@@ -1,6 +1,6 @@
 #include "framework.h"
 
-BEGIN_TEST(double)
+BEGIN_TEST(float)
 {
     c = (-(float)a);
 }
